+++ packages/patrol/lib/src/platform/android/android_automator_native.dart
@@ -56,6 +56,13 @@ class AndroidAutomator extends NativeMobileAutomator
 
   late final AndroidAutomatorClient _client;
 
+  @override
+  ConfigureRequest buildConfigureRequest() => ConfigureRequest(
+    findTimeoutMillis: _config.findTimeout.inMilliseconds,
+    androidDontSuppressAccessibilityServices:
+        _config.dontSuppressAccessibilityServices,
+  );
+
   /// Opens a platform-specific app.
   ///
   /// On Android, opens the app specified by [androidAppId] (package name).
@@ -536,6 +543,15 @@ class AndroidAutomator extends NativeMobileAutomator
           doneButtonSelector: doneButtonSelector,
           timeoutMillis: timeout?.inMilliseconds,
         ),
+      );
+    });
+  }
+
+  @override
+  Future<void> takeNativeScreenshot(String tag) async {
+    await wrapRequest('takeNativeScreenshot', () async {
+      await _client.takeNativeScreenshot(
+        AndroidTakeNativeScreenshotRequest(tag: tag),
       );
     });
   }